FUSETOOLS2-2104 - provide command to create SpringBoot project
it provides same level of functionality than to create Camel Quarkus
command

- validation of gav partial

It is picking all Camel files in the root workspace and copying them in
the src/main/resources:camel folder. if there are none, the structure is
still created and it is up to the users to create it.

I think will keep for another iteration:
- allow to open in another workspace the created project
- a more complete validation

Signed-off-by: Aurélien Pupier <apupier@redhat.com>